🔧 Fix Netlify build issues
✅ Fixed static export configuration for dynamic routes
✅ Separated server and client components properly
✅ Added generateStaticParams for driver pages
✅ Static export now generates 25 pages successfully
✅ Ready for Netlify deployment

Changes:
- Split dynamic route into server/client components
- Enabled static export with proper driver numbers
- Fixed Next.js 14 compatibility issues
- All builds now pass successfully
